+++ library/kani/Cargo.toml
@@ -10,6 +10,7 @@ publish = false
 
 [dependencies]
 kani_macros = { path = "../kani_macros" }
+kani_core = { path = "../kani_core" }
 
 [features]
 concrete_playback = []
